refactor: 인터페이스에서 구체클래스로 변경
- 현재 구조에서 인터페이스로 바꿀 필요는 없다고 판단
