+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,15 +1,38 @@
 package roomescape.service;
 
+import org.springframework.stereotype.Service;
 import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
 import roomescape.dto.ReservationRequest;
+import roomescape.repository.ReservationRepository;
 
 import java.util.List;
 
-public interface ReservationService {
+@Service
+public class ReservationService {
+    private final ReservationRepository reservationRepository;
+    private final TimeService timeService;
 
-    Reservation add(ReservationRequest request);
+    public ReservationService(ReservationRepository reservationRepository, TimeService timeService) {
+        this.reservationRepository = reservationRepository;
+        this.timeService = timeService;
+    }
 
-    List<Reservation> findAllReservations();
 
-    void deleteReservation(Long id);
+    public Reservation add(ReservationRequest request) {
+        ReservationTime time = timeService.findById(request.getTimeId());
+
+        Reservation reservation = new Reservation(request.getName(), request.getDate(), time);
+        return reservationRepository.add(reservation);
+    }
+
+
+    public List<Reservation> findAllReservations() {
+        return reservationRepository.findAllReservations();
+    }
+
+
+    public void deleteReservation(Long id) {
+        reservationRepository.remove(id);
+    }
 }
